refactor: cleaning up the trie iteration benchmark code
